fix handling of redirects on dashboard pages
